backend/requirements.txt
accelerate==1.2.1
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.16.4
annotated-types==0.6.0
ansicon==1.89.0
anthropic==0.3.6
anyio==4.1.0
appdirs==1.4.4
appnope==0.1.4
arabic-reshaper==3.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asn1crypto==1.5.1
astor==0.8.1
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.2
attr==0.3.2
attrs==23.1.0
Babel==2.14.0
backoff==2.2.1
beautifulsoup4==4.12.2
bert-score==0.3.13
bitsandbytes==0.43.3
bleach==6.1.0
blessed==1.20.0
blinker==1.9.0
blis==0.7.11
Brotli==1.1.0
camel-converter==3.0.2
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.15.1
colorama==0.4.6
comm==0.2.2
confection==0.1.3
contourpy==1.2.0
cryptography==45.0.3
cssselect==1.2.0
cssselect2==0.8.0
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.3
datasets==2.20.0
dateparser==1.2.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
distro==1.8.0
Django==4.2.5
docker-pycreds==0.4.0
docstring_parser==0.16
docx2txt==0.8
emoji==2.9.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
faiss-cpu==1.7.4
fastai==2.7.12
fastcore==1.5.29
fastdownload==0.0.7
fastjsonschema==2.19.0
fastprogress==1.0.3
feedfinder2==0.0.4
feedparser==6.0.11
ffmpeg==1.4
filelock==3.13.1
filetype==1.2.0
fire==0.4.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
fonttools==4.46.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.12.2
git-python==1.0.3
gitdb==4.0.10
GitPython==3.1.37
google-search-results==2.4.2
GoogleNews==1.6.11
greenlet==2.0.2
grpcio==1.58.0
grpcio-tools==1.58.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html5lib==1.1
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.33.4
hyperframe==6.0.1
idna==3.6
importlib-metadata==6.8.0
importlib-resources==6.1.1
iniconfig==2.0.0
inquirer==3.1.3
intel-openmp==2021.4.0
ipykernel==6.29.4
ipython==8.23.0
ipywidgets==8.1.5
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
jieba3k==0.35.1
Jinja2==3.1.2
jinxed==1.2.0
jiter==0.8.2
joblib==1.3.2
json5==0.9.14
jsonlines==4.0.0
jsonpatch==1.33
jsonpickle==3.0.2
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-spec==0.2.4
jsonschema-specifications==2023.11.2
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.12.1
jupyter_server_terminals==0.5.0
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==2.11.0
kiwisolver==1.4.5
krippendorff==0.7.0
lancedb==0.1.16
langchain==0.0.350
langchain-community==0.0.2
langchain-core==0.1.0
langchain-experimental==0.0.53
langchain-openai==0.0.5
langchain-text-splitters==0.0.1
langchainplus-sdk==0.0.20
langcodes==3.3.0
langdetect==1.0.9
langsmith==0.0.69
lazy-object-proxy==1.9.0
libcst==1.0.1
litellm==0.1.743
loguru==0.6.0
lxml==6.0.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.9.2
matplotlib-inline==0.1.6
mdurl==0.1.2
meilisearch==0.21.0
mistune==3.0.2
mkl==2021.4.0
more-itertools==10.1.0
moverscore==1.0.3
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.16
murmurhash==1.0.10
mypy-extensions==1.0.0
mysqlclient==2.2.7
nbclient==0.9.0
nbconvert==7.12.0
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
newspaper3k==0.2.8
nltk==3.8.1
notebook==7.2.2
notebook_shim==0.2.3
numexpr==2.8.6
numpy==1.26.2
open-interpreter==0.1.4
openai==1.40.0
openapi-core==0.18.1
openapi-schema-pydantic==1.2.4
openapi-schema-validator==0.6.1
openapi-spec-validator==0.6.0
openpyxl==3.0.10
orjson==3.10.0
oscrypto==1.3.0
overrides==7.4.0
packaging==23.2
pandas==1.5.3
pandocfilters==1.5.0
parse==1.19.1
parso==0.8.4
pathable==0.4.3
pathy==0.10.2
patsy==0.5.3
peft==0.12.0
pexpect==4.9.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==4.2.0
pluggy==1.3.0
portalocker==2.8.2
prance==23.6.21.0
preshed==3.0.9
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==6.31.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyarrow==18.1.0
pyarrow-hotfix==0.6
pyBKT==1.4.1
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pydyf==0.11.0
pyemd==1.0.0
Pygments==2.16.1
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
pykt-toolkit==0.0.38
pylance==0.5.10
PyMySQL==1.0.2
pypandoc==1.15
pyparsing==3.1.1
pypdf==3.17.2
pyphen==0.17.2
pyreadline3==3.4.1
pytest==7.2.2
pytest-mock==3.11.1
python-bidi==0.6.6
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-editor==1.0.4
python-iso639==2023.12.11
python-json-logger==2.0.7
python-magic==0.4.27
pytz==2023.3.post1
pyvis==0.3.2
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==25.1.2
qdrant-client==1.4.0
rapidfuzz==3.5.2
ratelimiter==1.2.0.post0
readchar==4.0.5
referencing==0.32.0
regex==2023.10.3
reportlab==4.4.3
requests==2.32.3
requests-file==1.5.1
retry==0.9.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.3
rouge==1.0.1
rpds-py==0.13.2
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
safetensors==0.4.5
scikit-learn==1.5.1
scipy==1.13.0
seaborn==0.13.0
semver==3.0.1
Send2Trash==1.8.2
sentence-transformers==3.1.0
sentencepiece==0.2.0
sentry-sdk==2.19.2
setproctitle==1.3.4
sgmllib3k==1.0.0
shtab==1.7.1
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.23
sqlparse==0.4.4
srsly==2.4.8
stack-data==0.6.3
statsmodels==0.13.5
svglib==1.5.1
sympy==1.14.0
ta==0.10.2
tabulate==0.9.0
tbb==2021.13.1
tenacity==8.2.3
termcolor==2.3.0
terminado==0.18.0
thinc==8.2.1
threadpoolctl==3.1.0
tiktoken==0.5.2
tinycss2==1.4.0
tinydb==4.8.2
tinyhtml5==2.0.0
tinysegmenter==0.3
tldextract==5.1.1
tokenizers==0.21.2
tokentrim==0.1.10
torch==2.7.1+cpu
torchaudio==2.7.1+cpu
torchvision==0.22.1
tornado==6.4
tqdm==4.66.5
traitlets==5.14.2
transformers==4.53.2
trl==0.9.6
typeguard==4.4.1
typer==0.7.0
types-python-dateutil==2.8.19.14
typing==3.7.4.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tyro==0.9.3
tzdata==2023.3
tzlocal==5.2
unstructured==0.11.2
uri-template==1.3.0
uritools==5.0.0
urllib3==2.1.0
wandb==0.17.0
wasabi==1.1.2
wcwidth==0.2.6
weasel==0.3.1
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.1.3
wget==3.2
widgetsnbextension==4.0.13
wikipedia==1.4.0
win32-setctime==1.1.0
wordcloud==1.9.3
wrapt==1.16.0
xhtml2pdf==0.2.17
XlsxWriter==3.0.8
xxhash==3.5.0
yachalk==0.1.5
yarl==1.9.4
zipp==3.17.0
zopfli==0.2.3.post1
